tx.in sequence to uint

+++ src/test/scala/donmiguel/tx/TxInSpec.scala
@@ -29,7 +29,7 @@ class TxInSpec extends UnitSpec {
 
     var txIn = TxIn.parse(CryptoUtil.hexToBytes(raw).iterator)
 
-    assert(txIn.sequence == 4294967294L)
+    assert(txIn.sequence.longValue() == 4294967294L)
   }
 
   it should "input value" in {
